+++ libtitania-x3d/Titania/X3D/Components/Geospatial/GeoElevationGrid.h
@@ -274,7 +274,7 @@ private:
 	createTexCoord () const;
 
 	std::vector <Vector3f>
-	createNormals (const std::vector <Vector3d> &, const std::vector <size_t> &, const float) const;
+	createNormals (const std::vector <Vector3d> &, const std::vector <size_t> &, const double) const;
 
 	std::vector <size_t>
 	createCoordIndex () const;
